Update to dor-services-client 1.0.0
In so doing, rescue the NotFoundResponse instead of UnexpectedResponse. The former is thrown when an object is not found in DOR, which we would expect for an object that has not yet been registered. This allows truly unexpected responses to raise an un-rescued exception, causing a stack trace (which we want in this case to make us aware of error conditions we have not yet anticipated).

Fixes #50
Refs sul-dlss/dor-services-client#38