Let fetch-files run
Depends on https://github.com/sul-dlss/common-accessioning/pull/1222
being merged.

After the ocrWF starts it should go get files in need of accessioning
from Preservation. If this step fails for whatever reason it can be
safely rerun since the start of the Abbyy job is done seprately, in the
next step start-ocr, or some subsequent yet-to-be-named step.

Closes #767